@@ -1,21 +1,13 @@
 {
   "name": "myapp",
   "author": "Ashish Nandan Singh",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@octokit/core": "^3.1.2",
-    "express": "^4.16.4",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
-    "react-ga": "^2.5.6",
-    "react-materialize": "^2.6.0",
-    "react-particles-js": "^2.4.2",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "^4.0.3",
-    "react-tweet": "^1.7.1",
-    "react-twitter-widgets": "^1.7.1",
-    "react-typing-animation": "^1.5.1"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.1",
+    "react-scripts": "^5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
