@@ -1,7 +1,10 @@
 import React, {Component} from "react"
 import "./App.css"
-import {Particles} from "react-particles-js"
-import {BrowserRouter as Router, Route} from "react-router-dom"
+import {
+  BrowserRouter as Router,
+  Routes,
+  Route,
+} from "react-router-dom";
 import Navigation from "./Components/Navigation/Navigation"
 import Home from "./Components/Pages/Home"
 import Projects from "./Components/Pages/Projects"
@@ -10,8 +13,6 @@ import Skills from "./Components/Pages/Skills"
 import Contact from "./Components/Pages/Contact"
 import TwitterFeed from "./Components/Pages/TwitterFeed"
 
-// initialise google analytics
-import ReactGA from "react-ga"
 
 class App extends Component {
   constructor(props) {
@@ -23,23 +24,20 @@ class App extends Component {
   }
 
   render() {
-    const {paramType} = this.state
-
-    ReactGA.initialize("UA-000000-01")
-    ReactGA.pageview(window.location.pathname + window.location.search)
-
     return (
       <Router>
         <>
           <div className="App">
-            <Particles className="background" params={paramType} />
-            <Navigation />
-            <Route exact path="/" component={Home} />
-            <Route exact path="/projects" component={Projects} />
-            <Route exact path="/blog" component={Blog} />
-            <Route exact path="/skills" component={Skills} />
-            <Route exact path="/contact" component={Contact} />
-            <Route exact path="/twitter" component={TwitterFeed} />
+            {/* <Navigation /> */}
+            <Home/>
+            <Routes>
+            <Route  path="/" element={Home} />
+            <Route  path="/projects" element={Projects} />
+            <Route  path="/blog" element={Blog} />
+            <Route  path="/skills" element={Skills} />
+            <Route  path="/contact" element={Contact} />
+            <Route  path="/twitter" element={TwitterFeed} />
+            </Routes>
           </div>
         </>
       </Router>
